requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-storage==2.10.0
Werkzeug==2.3.7
google-generativeai==0.8.2
